requirements.txt
Django==4.2.14        # LTS toujours maintenue
djongo==1.3.6         # stable avec Django 4.2
pymongo==4.7
python-dotenv==1.0
pandas==2.2
scikit-learn==1.5
scipy==1.13
